#1049 P1: relocate filter.rs inline tests to filter_tests.rs
filter.rs went from 2,278 raw LOC (1,414 prod / 862 tests) to 1,419
by moving the bottom `#[cfg(test)] mod tests { ... }` block (line
1417) into a sibling `filter_tests.rs` file, loaded via
`#[cfg(test)] #[path = "filter_tests.rs"] mod tests;` appended to
filter.rs.

Note: filter.rs has TWO smaller `#[cfg(test)]` decorations on
production helper fns (`record_filter_counter` at line 173 and the
test-only stub of `flush_recorded_filter_counters` at line 185).
Those stay in place — they are file-level test-gated production
helpers, not part of the `mod tests` block.

Pure file-organization split — no test bodies modified, no production
code touched. Test module path `crate::filter::tests` is unchanged.

LOC distribution after:
- filter.rs:        1,419 (was 2,278) — under threshold
- filter_tests.rs:    863 (new — same content, dedented one level)

cargo test --release: 865 passed, 0 failed.
